Update ServiceNowRequest.cs
Removal of the username and password values on line 38, 39 and setting them to an empty string value. Huu Ly has confirmed that this API code is no longer in use, but would like to keep the code in Github for historical purposes. This should no longer violate any Github security concerns.